+++ changelog.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+### features
+- allow change default output volume using mouse wheel on media module items.
+
 ### enhancements
 - add new settings to delay the show and hide of dock and toolbar.